refactor: GameCardStatus型をGameCard型に変更
