+++ src/features/management/components/StatusGameCard.tsx
@@ -2,9 +2,10 @@ import { Button, Card, Group, Image, Text } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
 import StatusUpdateModal from "features/status/components/StatusUpdateModal";
 import { useMediaQuery } from "lib/mantine/useMediaQuery";
-import { GameCardStatus } from "types/game";
 
-const StatusGameCard = ({ gameData }: { gameData: GameCardStatus }) => {
+import { GameCard } from "../types";
+
+const StatusGameCard = ({ gameData }: { gameData: GameCard }) => {
   const largerThanSm = useMediaQuery("sm");
   const [opened, { open, close }] = useDisclosure(false);
